requirements.txt
torch
transformers
chromadb
langchain
sentence_transformers
chromadb
langchain-community
langchain-huggingface
PyMuPDF
arxiv
playwright
requests
tweepy
python-dotenv
ollama
arxiv
Flask
requests
sentencepiece
accelerate
pypdf2
psutil
lxml
datasets
peft